refactor(pos): extrae SelectorDeLote a un componente compartido
Ningun cambio de comportamiento — mismo componente, movido de un
local no exportado dentro de Pos.tsx a componentes/SelectorDeLote.tsx.
Prepara la reutilizacion literal que design.md pide para Remito.tsx
("lot picker reusing SelectorDeLote"). Pos.test.tsx sigue verde sin
tocarse (nunca importo el componente directamente).
